Refactored redis cache #621.
