+++ src/Abp.RedisCache/RedisCache/AbpRedisCacheManager.cs
@@ -1,17 +1,11 @@
 ﻿using Abp.Dependency;
 using Abp.Runtime.Caching;
 using Abp.Runtime.Caching.Configuration;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Abp.RedisCache
 {
     public class AbpRedisCacheManager : CacheManagerBase
     {
-
         public AbpRedisCacheManager(IIocManager iocManager, ICachingConfiguration configuration)
             : base(iocManager, configuration)
         {
