+++ src/Abp.RedisCache/RedisCache/Configuration/IRedisConnectionProvider.cs
@@ -1,15 +1,11 @@
 ﻿using StackExchange.Redis;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Abp.RedisCache
+namespace Abp.RedisCache.Configuration
 {
     public interface IAbpRedisConnectionProvider 
     {
         ConnectionMultiplexer GetConnection(string connectionString);
+
         string GetConnectionString(string service);
     }
 }
