+++ src/Abp.RedisCache/RedisCache/RedisImpl/RedisDatabaseExtensions.cs
@@ -1,15 +1,10 @@
 ﻿using StackExchange.Redis;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Abp.RedisCache.RedisImpl
 {
     public static class RedisDatabaseExtensions
     {
-
         public static void KeyDeleteWithPrefix(this IDatabase database, string prefix)
         {
             if (database == null)
